+++ kinto/core/views/hello.py
@@ -48,6 +48,10 @@ def get_hello(request):
     # (Note: this will call authenticated_userid() with multiauth+groupfinder)
     if Authenticated in request.effective_principals:
         data["user"] = request.get_user_info()
+    else:
+        # Information can be cached.
+        cache_seconds = int(settings["root_cache_expires_seconds"])
+        request.response.cache_expires(cache_seconds)
 
     # Application can register and expose arbitrary capabilities.
     data["capabilities"] = request.registry.api_capabilities