+++ tests/test_views_hello.py
@@ -22,6 +22,15 @@ class HelloViewTest(BaseWebTest, unittest.TestCase):
         response = self.app.get("/")
         self.assertNotIn("user", response.json)
 
+    def test_cache_control_if_anonymous(self):
+        response = self.app.get("/")
+        self.assertIn("Cache-Control", response.headers)
+        self.assertEqual(response.headers["Cache-Control"], "max-age=86400")
+
+    def test_no_cache_control_if_authenticated(self):
+        response = self.app.get("/", headers=self.headers)
+        self.assertNotIn("Cache-Control", response.headers)
+
     def test_returns_user_id_if_authenticated(self):
         response = self.app.get("/", headers=self.headers)
         self.assertEqual(response.json["user"]["id"], self.principal)